Add files via upload
Hi. so yes it is currently a .ps1 script because most people probably have windows. Yes it is slow at the moment because i keep adding to it. 2 things should ALWAYS be true about this script: 1. It will either be 100% helpful to any bug hunter or 0% helpful. 2. It will be endlessly updated unless they stop putting out new endpoint header security features and mitigations. I really don't care it the script is a bash script, powershell script, python script, "go install" tool. Whatever! as long as: it is easy to update/add to and IT WORKS AS INTENDED! THIS IS ALL I CARE ABOUT!