Add files via upload
My suggestion: If you are current a bug hunter: put this .ps1 script (Windows 11 Powershell) in the same directory where you have a urlslist.txt file: run the script in powershell (.\web2endpointheaderfactscanner.ps1): choose option 3 (your urlslist.txt): name that .txt file: hit enter and see what it does. Yes it is slow at the moment because i keep adding for for the script to check for.